+++ src/components/NpcGenerator.vue
@@ -1,0 +1,24 @@
+<template>
+    <div class="container">
+      </div>
+</template>
+
+<script>
+
+    export default {
+        data() {
+            return {
+            }
+        },
+        methods: {
+        },
+
+        computed: {},
+
+        mounted() {
+        }
+    }
+</script>
+
+<style scoped>
+</style>