+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Fix Herbalism table column order and width
+- Npc Generator added
 
 ## 0.3.0
 - Add style to route if on route